requirements.txt
# Core
torch==2.1.0
transformers==4.36.0
datasets==2.16.0
accelerate==0.25.0
sentence-transformers==2.2.2

# Retrieval
rank-bm25==0.2.2
elasticsearch==8.11.0

# Evaluation
evaluate==0.4.1
rouge-score==0.1.2
bert-score==0.3.13
nltk==3.8.1
sacrebleu==2.3.1

# Utilities
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.0

# Experiment tracking
wandb==0.16.0
tensorboard==2.15.1

# Development
pytest==7.4.3
black==23.11.0
jupyterlab==4.0.8
ipykernel==6.27.1

# LLM optimization
bitsandbytes==0.41.3
peft==0.7.0
trl==0.7.4

# HuggingFace
huggingface-hub==0.20.2